+++ integrated-scouting-apps/src/public/stylesheets/style.css
@@ -26,7 +26,7 @@ h2 {
 body {
     margin: 0;
     font-family: Audiowide !important;
-    background-color: #32a7dc;
+    background-color: #433C3C;
     padding-left: 5%;
     padding-right: 5%;
     color: white;
@@ -46,15 +46,8 @@ a {
     height: 125px !important;
 }
 
-.ant-tabs-tab.ant-tabs-tab-active .ant-tabs-tab-btn {
-    color: black !important;
-}
-.ant-tabs-ink-bar {
-    background-color: black !important;
-}
-
 .ant-tabs-ink-bar-animated {
-    transition: width 0.2s,left 0.2s,right 0.2s !important;
+    transition: width 0.2s, left 0.2s, right 0.2s !important;
 }
 
 .banner {
@@ -66,8 +59,8 @@ a {
 .mainbutton:hover {
     width: 100%;
     min-width: 100%;
-    height: 20%;
-    min-height: 20%;
+    height: 25% !important;
+    min-height: 25% !important;
     margin-top: 10%;
     border: 4px solid white !important;
     border-radius: 25px;
