+++ integrated-scouting-apps/src/routes/picklists.tsx
@@ -7,7 +7,6 @@ import VerifyLogin from '../verifyToken';
 import { useCookies } from 'react-cookie';
 import Column from 'antd/es/table/Column';
 import { Table } from 'antd';
-import { saveAs } from 'file-saver';
 
 function Picklists(props: any) {
   const [cookies] = useCookies(['login', 'theme']);
@@ -21,10 +20,9 @@ function Picklists(props: any) {
       try {
         setLoading(true);
         const teams: { [key: string]: any } = {};
-        const kv = [];
+        let kv = [];
         const response = await fetch(process.env.REACT_APP_PICKLIST_URL as string);
         const data = await response.json();
-        saveAs(new Blob([JSON.stringify(data)], { type: "text/json" }), "j.json");
         for (const team in data) {
           if (teams[data[team].team_number] === undefined) {
             teams[data[team].team_number] = (data[team]);
@@ -40,8 +38,11 @@ function Picklists(props: any) {
             if (teams[data[team].team_number].robot_died === undefined) {
               teams[data[team].team_number].robot_died = 0;
             }
-            if (teams[data[team].team_number].match_number === undefined) {
-              teams[data[team].team_number].match_number = 0;
+            if (teams[data[team].team_number].match_numbers === undefined) {
+              teams[data[team].team_number].match_numbers = 1;
+            }
+            if (teams[data[team].team_number].rank === undefined) {
+              teams[data[team].team_number].rank = 1;
             }
           }
           else {
@@ -66,29 +67,38 @@ function Picklists(props: any) {
             teams[data[team].team_number].teleop_missed_amps += data[team].teleop_missed_amps;
             teams[data[team].team_number].teleop_missed_speaker += data[team].teleop_missed_speaker;
 
-            teams[data[team].team_number].match_number++;
+            teams[data[team].team_number].match_numbers++;
+            // console.log(teams[data[team].team_number].team_number + ": first_score: " + teams[data[team].team_number].first_score);
+            // console.log(teams[data[team].team_number].team_number + ": second_score: " + teams[data[team].team_number].second_score);
+            // console.log(teams[data[team].team_number].team_number + ": third_score: " + teams[data[team].team_number].third_score);
+            // console.log(teams[data[team].team_number].team_number + ": match_numbers: " + teams[data[team].team_number].match_numbers);
+            // console.log(teams[data[team].team_number].team_number + ": score: " + teams[data[team].team_number].score);
           }
           if (data[team].robot_died) {
             teams[data[team].team_number].robot_died++;
           }
         }
         for (const team in teams) {
-          teams[teams[team].team_number].avg_score = Math.round((teams[teams[team].team_number].score / teams[teams[team].team_number].match_number) * 100) / 100;
-          teams[teams[team].team_number].iegr = Math.round((teams[teams[team].team_number].first_score + teams[teams[team].team_number].second_score + teams[teams[team].team_number].third_score) / 3 * 100) / 100;
+          teams[teams[team].team_number].avg_score = Math.round(teams[teams[team].team_number].score / teams[teams[team].team_number].match_numbers * 100) / 100;
+          teams[teams[team].team_number].iegr = Math.round(((teams[teams[team].team_number].first_score + teams[teams[team].team_number].second_score + teams[teams[team].team_number].third_score) / 3) * 100) / 100;
 
           teams[teams[team].team_number].auto_score_ratio = teams[teams[team].team_number].auto_scored_pieces + "/" + (teams[teams[team].team_number].auto_scored_pieces + teams[teams[team].team_number].auto_missed_pieces);
-          teams[teams[team].team_number].auto_scored_pieces = Math.round(teams[teams[team].team_number].auto_scored_pieces / teams[teams[team].team_number].match_number * 100) / 100;
+          teams[teams[team].team_number].auto_scored_pieces = Math.round(teams[teams[team].team_number].auto_scored_pieces / teams[teams[team].team_number].match_numbers * 100) / 100;
 
           teams[teams[team].team_number].amp_score_ratio = teams[teams[team].team_number].teleop_amps_scored + "/" + (teams[teams[team].team_number].teleop_amps_scored + teams[teams[team].team_number].teleop_missed_amps);
           teams[teams[team].team_number].speaker_score_ratio = teams[teams[team].team_number].teleop_speaker_scored + "/" + (teams[teams[team].team_number].teleop_speaker_scored + teams[teams[team].team_number].teleop_missed_speaker);
-          teams[teams[team].team_number].teleop_amps_scored = Math.round(teams[teams[team].team_number].teleop_amps_scored / teams[teams[team].team_number].match_number * 100) / 100;
-          teams[teams[team].team_number].teleop_speaker_scored = Math.round(teams[teams[team].team_number].teleop_speaker_scored / teams[teams[team].team_number].match_number * 100) / 100;
+          teams[teams[team].team_number].teleop_amps_scored = Math.round(teams[teams[team].team_number].teleop_amps_scored / teams[teams[team].team_number].match_numbers * 100) / 100;
+          teams[teams[team].team_number].teleop_speaker_scored = Math.round(teams[teams[team].team_number].teleop_speaker_scored / teams[teams[team].team_number].match_numbers * 100) / 100;
         }
         for (const team in teams) {
           kv.push(teams[team]);
+
+        }
+        kv.sort((a, b) => {return (b.avg_score - a.avg_score);});
+        for (const team in kv) {
+          kv[team].rank = Number(team) + 1;
         }
         setFetchedData(kv);
-        console.log(kv)
       }
       catch (err) {
         console.log(err);
@@ -123,17 +133,17 @@ function Picklists(props: any) {
         </header>
         <h2 style={{ whiteSpace: 'pre-line' }}>{loading ? 'Loading Data...' : ''}</h2>
         <Table dataSource={fetchedData} pagination={{ pageSize: 500 }}>
-          <Column title="Rank #" dataIndex="rank" key="rank" sorter={(a: any, b: any) => a.rank - b.rank} />
-          <Column title="Team #" dataIndex="team_number" key="team_number" sorter={(a: any, b: any) => a.team_number - b.team_number} />
-          <Column title="Overall Score" dataIndex="avg_score" key="avg_score" sorter={(a: any, b: any) => a.avg_score - b.avg_score} defaultSortOrder={"descend"} />
-          <Column title="IEGR (Top 3)" dataIndex="iegr" key="iegr" sorter={(a: any, b: any) => a.iegr - b.iegr} />
-          <Column title="Auton Avg" dataIndex="auto_scored_pieces" key="auto_scored_pieces" sorter={(a: any, b: any) => a.auto_scored_pieces - b.auto_scored_pieces} />
-          <Column title="Auton (Scr/Att)" dataIndex="auto_score_ratio" key="auto_score_ratio" sorter={(a: any, b: any) => a.auto_score_ratio.length - b.auto_score_ratio.length} />
-          <Column title="Teleop Amp Avg" dataIndex="teleop_amps_scored" key="teleop_amps_scored" sorter={(a: any, b: any) => a.teleop_amps_scored - b.teleop_amps_scored} />
-          <Column title="Teleop Amp (Scr/Att)" dataIndex="amp_score_ratio" key="amp_score_ratio" sorter={(a: any, b: any) => a.amp_score_ratio.length - b.amp_score_ratio.length} />
-          <Column title="Teleop Spkr Avg" dataIndex="teleop_speaker_scored" key="teleop_speaker_scored" sorter={(a: any, b: any) => a.teleop_speaker_scored - b.teleop_speaker_scored} />
-          <Column title="Teleop Spkr (Scr/Att)" dataIndex="speaker_score_ratio" key="speaker_score_ratio" sorter={(a: any, b: any) => a.speaker_score_ratio.length - b.speaker_score_ratio.length} />
-          <Column title="Robot Deaths" dataIndex="robot_died" key="robot_died" sorter={(a: any, b: any) => a.robot_died - b.robot_died} className='robotdied' />
+          <Column title="Rank #" dataIndex="rank" sorter={(a: any, b: any) => a.rank - b.rank} defaultSortOrder={"ascend"} fixed="left" />
+          <Column title="Team #" dataIndex="team_number" sorter={(a: any, b: any) => a.team_number - b.team_number} />
+          <Column title="Overall Score" dataIndex="avg_score" sorter={(a: any, b: any) => a.avg_score - b.avg_score} />
+          <Column title="IEGR (Top 3)" dataIndex="iegr" sorter={(a: any, b: any) => a.iegr - b.iegr} />
+          <Column title="Auton Avg" dataIndex="auto_scored_pieces" sorter={(a: any, b: any) => a.auto_scored_pieces - b.auto_scored_pieces} />
+          <Column title="Auton (Scr/Att)" dataIndex="auto_score_ratio" sorter={(a: any, b: any) => a.auto_score_ratio.length - b.auto_score_ratio.length} />
+          <Column title="Teleop Amp Avg" dataIndex="teleop_amps_scored" sorter={(a: any, b: any) => a.teleop_amps_scored - b.teleop_amps_scored} />
+          <Column title="Teleop Amp (Scr/Att)" dataIndex="amp_score_ratio" sorter={(a: any, b: any) => a.amp_score_ratio.length - b.amp_score_ratio.length} />
+          <Column title="Teleop Spkr Avg" dataIndex="teleop_speaker_scored" sorter={(a: any, b: any) => a.teleop_speaker_scored - b.teleop_speaker_scored} />
+          <Column title="Teleop Spkr (Scr/Att)" dataIndex="speaker_score_ratio" sorter={(a: any, b: any) => a.speaker_score_ratio.length - b.speaker_score_ratio.length} />
+          <Column title="Robot Deaths" dataIndex="robot_died" sorter={(a: any, b: any) => a.robot_died - b.robot_died} className='robotdied' />
         </Table>
       </div>
     </div>
